# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doctor Management & Appointment Booking API",
        "description": "REST API for the Doctor Management and Simple Appointment Booking System.\n\nRoles: Management (admin), Doctor, Regular user (patient).\n\nAuthentication uses Laravel Sanctum bearer tokens. Obtain a token from POST /api/v1/login, then send it as `Authorization: Bearer <token>`.",
        "contact": {
            "name": "Support",
            "email": "mebrahimi405@yahoo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development"
        }
    ],
    "paths": {
...
        "/api/v1/doctor/appointment": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "summary": "View own appointments",
                "description": "Display a listing of the resource.",
                "operationId": "doctorListAppointments",
                "responses": {
                    "200": {
                        "description": "List of appointments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Appointment"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/doctor/appointment/{appointment}": {
            "patch": {
                "tags": [
                    "Doctor"
                ],
                "summary": "Approve / cancel an appointment",
                "description": "Change an appointment's status. Use 'approved' to approve or 'canceled' to cancel.",
                "operationId": "doctorUpdateAppointmentStatus",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "description": "Appointment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAppointmentStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated appointment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Appointment"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/doctor/profile/{profile}": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "summary": "View own doctor profile",
                "operationId": "doctorShowProfile",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "Doctor profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor profile with user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoctorProfileDetail"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Doctor"
                ],
                "summary": "Complete / update own profile",
                "operationId": "doctorUpdateProfile",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "Doctor profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDoctorProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated doctor profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoctorProfileDetail"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/patient/appointment": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "summary": "View approved (active) doctors",
                "operationId": "patientListActiveDoctors",
                "responses": {
                    "200": {
                        "description": "List of active doctors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DoctorProfileFull"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Patient"
                ],
                "summary": "Book an appointment",
                "operationId": "patientBookAppointment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatientAppointmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment booked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookAppointmentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "DoctorStatus": {
                "description": "Doctor profile status",
                "type": "string",
                "enum": [
                    "pending",
                    "active",
                    "rejected",
                    "suspended",
                    "vacation",
                    "archived"
                ]
            },
            "AppointmentStatus": {
                "description": "Appointment status",
                "type": "string",
                "enum": [
                    "pending",
                    "approved",
                    "canceled"
                ]
            },
...
            "PatientAppointmentRequest": {
                "required": [
                    "doctor_id",
                    "appointment_date",
                    "appointment_time"
                ],
                "properties": {
                    "doctor_id": {
                        "description": "Doctor profile ID",
                        "type": "integer",
                        "example": 1
                    },
                    "appointment_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-08-20"
                    },
                    "appointment_time": {
                        "type": "string",
                        "example": "14:30"
                    },
                    "description": {
                        "type": "string",
                        "example": "Sore throat and fever",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateAppointmentStatusRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "$ref": "#/components/schemas/AppointmentStatus"
                    }
                },
                "type": "object"
            },
            "UpdateDoctorProfileRequest": {
                "properties": {
                    "specialty_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "image": {
                        "description": "Image path or URL",
                        "type": "string",
                        "example": "doctors/avatar.jpg"
                    },
                    "bio": {
                        "type": "string",
                        "example": "Experienced cardiologist with 10+ years."
                    },
                    "mobile": {
                        "type": "string",
                        "example": "+989123456789"
                    },
                    "medical_code": {
                        "type": "string",
                        "example": "MC-12345"
                    },
                    "address": {
                        "type": "string",
                        "example": "Tehran, Valiasr St."
                    },
                    "working_hours": {
                        "type": "string",
                        "example": "Sat-Wed 9:00-17:00"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin User"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@test.com"
                    }
                },
                "type": "object"
            },
...
            "DoctorProfileFull": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "specialty_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "status": {
                        "$ref": "#/components/schemas/DoctorStatus"
                    },
                    "image": {
                        "type": "string",
                        "example": "doctors/avatar.jpg",
                        "nullable": true
                    },
                    "bio": {
                        "type": "string",
                        "example": "Experienced cardiologist.",
                        "nullable": true
                    },
                    "mobile": {
                        "type": "string",
                        "example": "+989123456789",
                        "nullable": true
                    },
                    "medical_code": {
                        "type": "string",
                        "example": "MC-12345",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "Tehran, Valiasr St.",
                        "nullable": true
                    },
                    "working_hours": {
                        "type": "string",
                        "example": "Sat-Wed 9:00-17:00",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "DoctorProfileDetail": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/DoctorProfileFull"
                    },
                    {
                        "properties": {
                            "user": {
                                "$ref": "#/components/schemas/User"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "Appointment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "Patient user ID",
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "doctor_id": {
                        "description": "Doctor profile ID",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "appointment_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-08-20"
                    },
                    "appointment_time": {
                        "type": "string",
                        "example": "14:30:00"
                    },
                    "description": {
                        "type": "string",
                        "example": "Sore throat and fever",
                        "nullable": true
                    },
                    "status": {
                        "$ref": "#/components/schemas/AppointmentStatus"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "BookAppointmentResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Appointment booked successfully."
                    },
                    "data": {
                        "$ref": "#/components/schemas/Appointment"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "Unauthorized": {
                "description": "Unauthenticated - missing or invalid token.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Forbidden - the authenticated user does not have the required role/permission.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthorized action."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Resource not found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "No query results for model."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ValidationError": {
                "description": "Validation failed.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "The given data was invalid."
                                },
                                "errors": {
                                    "type": "object",
                                    "example": {
                                        "email": [
                                            "The email field is required."
                                        ]
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token. Enter the token returned by /api/v1/login in the format \"Bearer <token>\".",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}